Bug 30719: (QA follow-up) Fix tiny boolean is_system

+++ installer/data/mysql/kohastructure.sql
@@ -3290,7 +3290,7 @@ CREATE TABLE `illbatch_statuses` (
     `id` int(11) NOT NULL auto_increment COMMENT "Status ID",
     `name` varchar(100) NOT NULL COMMENT "Name of status",
     `code` varchar(20) NOT NULL COMMENT "Unique, immutable code for status",
-    `is_system` int(1) COMMENT "Is this status required for system operation",
+    `is_system` tinyint(1) COMMENT "Is this status required for system operation",
     PRIMARY KEY (`id`),
     UNIQUE KEY `u_illbatchstatuses__code` (`code`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
